requirements.txt
annotated-types==0.6.0
anyio==4.0.0
appdirs==1.4.4
argh==0.26.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
async-lru==2.0.4
atomicwrites==1.4.0
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
backports.weakref==1.0.post1
beautifulsoup4==4.12.2
bkcharts==0.2
black==19.10b0
blackstone==0.1.14
bleach==6.1.0
blis==0.7.9
boto==2.49.0
boto3==1.28.66
botocore==1.31.66
Bottleneck==1.3.2
brotlipy==0.7.0
catalogue==1.0.2
certifi==2020.12.5
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudpathlib==0.16.0
clyent==1.2.1
comm==0.1.4
confection==0.1.3
conllu==4.5.1
contextlib2==0.6.0.post1
contourpy==1.1.1
cvxopt==1.3.0
cycler==0.10.0
cymem==2.0.6
cytoolz==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dpcpp-cpp-rt==2023.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0.tar.gz
en-blackstone-proto @ https://blackstone-model.s3-eu-west-1.amazonaws.com/en_blackstone_proto-0.0.1.tar.gz
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-2.3.0/en_core_web_lg-2.3.0.tar.gz
entrypoints==0.3
et-xmlfile==1.0.1
exceptiongroup==1.1.3
executing==2.0.0
fastcache==1.1.0
fastjsonschema==2.18.1
fonttools==4.43.1
fqdn==1.5.1
functions==0.7.0
future==0.18.2
gatenlp==1.0.7
graphviz==0.20.1
h5py==2.10.0
HeapDict==1.0.1
helpdev==0.7.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
intel-cmplr-lib-rt==2023.2.0
intel-cmplr-lic-rt==2023.2.0
intel-opencl-rt==2023.2.0
intel-openmp==2023.2.0
interchange==2021.0.4
intervaltree==3.1.0
ipykernel==6.25.2
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.1
isodate==0.6.1
isoduration==20.11.0
jdcal==1.4.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
json5==0.9.5
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyter_server==2.8.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
langcodes==3.3.0
llvmlite==0.36.0
locket==0.2.1
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mccabe==0.6.1
mistune==3.0.2
mkl==2023.2.0
mkl-fft==1.3.0
mkl-service==2.4.0
monotonic==1.6
mpld3==0.5.7
mpmath==1.2.1
multipledispatch==0.6.0
murmurhash==1.0.6
mypy-extensions==0.4.3
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
neo4j-driver==5.8.1
nest-asyncio==1.5.8
networkx==3.1
neuralcoref==4.0
normalization==0.4
notebook==7.0.6
notebook_shim==0.2.3
numexpr==2.8.6
numpy==1.24.4
olefile==0.46
overrides==7.4.0
packaging==23.2
pandas==1.2.4
pandocfilters==1.5.0
pansi==2020.7.3
parso==0.8.3
pathspec==0.11.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pkginfo==1.7.0
pkgutil_resolve_name==1.3.10
plac==1.1.3
platformdirs==3.11.0
pluggy==0.13.1
ply==3.11
pm4py==2.7.2
preshed==3.0.6
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.17.2
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py2neo==2021.2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydot==1.4.2
pydotplus==2.0.2
Pygments==2.16.1
pyparsing==3.1.1
pytest==6.2.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
QDarkStyle==2.8.1
qtconsole==5.4.4
QtPy==2.4.0
rdflib==6.3.2
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
s3transfer==0.7.0
scipy==1.10.1
Send2Trash==1.8.2
simplegeneric==0.8.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==2.3.9
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==1.0.7
stack-data==0.6.3
StringDist==1.0.9
tables==3.6.1
tbb==2021.10.0
terminado==0.9.4
thinc==7.4.6
tifffile==2020.10.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typed-ast==1.5.5
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
unicodecsv==0.14.1
uri-template==1.3.0
urllib3==1.26.18
wasabi==0.10.1
wcwidth==0.2.8
weasel==0.3.3
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
xlwt==1.3.0
zict==2.0.0
zipp==3.17.0
zope.event==4.5.0
